requirements.txt
#
fastapi==0.111.0
fastapi-login==1.10.1
Jinja2==3.1.4
uvicorn==0.30.1

python-dotenv==1.0.1
pydantic-settings==2.3.4
pydantic>=2.7.0

pytest==8.2.2
